# src/config/swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "Nahual API",
        "description": "Nahual API",
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        }
    },
    "host": "localhost:8000",
    "basePath": "/api",
...
    "schemes": [
        "http"
    ],
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
...
        "/estudiantes/{id}": {
            "get": {
                "tags": [
                    "Estudiantes"
                ],
                "summary": "Obtener toda la informacion de un estudiante",
                "description": "Obtener toda la informacion de un estudiante",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "description": "ID de un estudiante",
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Estudiante"
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Estudiantes"
                ],
                "summary": "Enviar la informacion de un estudiante",
                "description": "Enviar la informacion de un estudiante",
                "parameters": [
                    {
                        "name": "Estudiante a guardar",
                        "in": "body",
                        "description": "Toda la informacion de un estudiante",
                        "schema": {
                            "$ref": "#/definitions/EstudianteSinId"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "El estudiante fue creado exitosamente",
                        "schema": {
                            "$ref": "#/definitions/EstudianteSinId"
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Estudiantes"
                ],
                "summary": "Editar toda la informacion de un estudiante especifico",
                "description": "Editar toda la informacion de un estudiante especifico",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "description": "ID del estudiante",
                        "type": "integer"
                    },
                    {
                        "name": "estudiante",
                        "in": "body",
                        "description": "Toda la informacion actualizada de un estudiante",
                        "type": "string",
                        "schema": {
                            "$ref": "#/definitions/EstudianteSinId"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "La informacion del estudiante fue actualizada exitosamente",
                        "schema": {
                            "$ref": "#/definitions/EstudianteSinId"
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Estudiantes"
                ],
                "summary": "Elimina a un estudiante especifico",
                "description": "Elimina a un estudiante especifico",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "description": "ID del estudiante",
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Estudiante fue eliminado exitosamente."
                    }
                }
            }
        },
...
        "/cursos/{id}": {
            "get": {
                "tags": [
                    "Cursos"
                ],
                "summary": "Muestra un curso en especifico",
                "description": "Muestra un curso en especifico por el id del URL",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "description": "ID del curso",
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "El curso fue editado exitosamente",
                        "schema": {
                            "$ref": "#/definitions/Curso"
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Cursos"
                ],
                "summary": "Editar toda la informacion de un curso especifico",
                "description": "Editar toda la informacion de un curso especifico",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "description": "ID del curso",
                        "type": "integer"
                    },
                    {
                        "name": "Curso nuevo",
                        "in": "body",
                        "description": "Toda la informacion de un curso",
                        "schema": {
                            "$ref": "#/definitions/CursoEditar"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "El curso fue editado exitosamente",
                        "schema": {
                            "$ref": "#/definitions/Curso"
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Cursos"
                ],
                "summary": "Elimina un curso en especifico",
                "description": "Elimina un curso en especifco por el iD en el URL",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "description": "ID del curso",
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "El curso fue eliminado exitosamente"
                    }
                }
            }
        },
...
        "/topicos/{id}": {
            "get": {
                "tags" : [
                    "Topicos"
                ],
                "summary": "Obtiene el topico identificado por el id",
                "description": "Obtiene solo un topico con el mismo id que el URL",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id del topico",
                        "type": "integer"
                    }                    
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/DatoTopico"
                        }
                    },
                    "404":{
                        "description": "Topico no encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Topicos"
                ],
                "summary": "Modifica un topico ya existente",
                "description": "Busca el topico con el id del URL y lo modifica",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id del topico",
                        "type": "integer"
                    },
                    {
                        "name": "nombre",
                        "in": "body",
                        "description": "Nuevo nombre para el topico",
                        "type": "string",
                        "schema": {
                            "$ref": "#/definitions/PostTopico"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "El topico fue editado exitosamente",
                        "schema": {
                            "$ref": "#/definitions/DatoTopico"
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Topicos"
                ],
                "summary": "Elimina un topico en especifico",
                "description": "Busca el topico con el id del URL y lo elimina",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id del topico",
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "El topico fue eliminado exitosamente"
                    }
                }
                
            }
        }
    },   
    "definitions": {
...
        "Estudiante": {
            "properties": {
                "id": {
                    "type": "integer",
                    "uniqueItems": true
                },
                "dni": {
                    "type": "string"
                },
                "tipoDNI": {
                    "type": "string"
                },
                "nombre": {
                    "type": "string"
                },
                "apellido": {
                    "type": "string"
                },
                "nacionalidad": {
                    "type": "string"
                },
                "fechaNacimiento": {
                    "type": "string"
                },
                "correo": {
                    "type": "string"
                },
                "correoOpcional": {
                    "type": "string"
                },
                "celular": {
                    "type": "string"
                },
                "linkedin": {
                    "type": "string"
                },
                "zona": {
                    "type": "string"
                },
                "detalle": {
                    "type": "string"
                },
                "añoGraduacion": {
                    "type": "integer"
                },
                "nombrePrimerTrabajo": {
                    "type": "string"
                },
                "trabajaActualmente": {
                    "type": "boolean"
                },
                "trabajaSistemas": {
                    "type": "boolean"
                },
                "experienciaSistemas": {
                    "type": "string"
                },
                "estudiosSistemas": {
                    "type": "string"
                },
                "fechaActualTrabajo": {
                    "type": "string"
                },
                "lugarActualTrabajo": {
                    "type": "string"
                },
                "fechaPrimerEmpleo": {
                    "type": "string"
                },
                "topico": {
                    "type": "string"
                },
                "nivelInglesId": {
                    "type": "integer",
                    "$ref": "#/definitions/NivelIngles"
                },
                "estadoId": {
                    "type": "integer",
                    "$ref": "#/definitions/Estado"
                },
                "sedeId": {
                    "type": "integer",
                    "$ref": "#/definitions/Sede"
                },
                "nodoId": {
                    "type": "integer",
                    "$ref": "#/definitions/Nodo"
                }
            }
        },
        "EstudianteSinId": {
            "properties": {
                "id": {
                    "type": "integer",
                    "uniqueItems": true
                },
                "dni": {
                    "type": "string"
                },
                "tipoDNI": {
                    "type": "string"
                },
                "nombre": {
                    "type": "string"
                },
                "apellido": {
                    "type": "string"
                },
                "nacionalidad": {
                    "type": "string"
                },
                "fechaNacimiento": {
                    "type": "string"
                },
                "correo": {
                    "type": "string"
                },
                "correoOpcional": {
                    "type": "string"
                },
                "celular": {
                    "type": "string"
                },
                "linkedin": {
                    "type": "string"
                },
                "zona": {
                    "type": "string"
                },
                "detalle": {
                    "type": "string"
                },
                "añoGraduacion": {
                    "type": "integer"
                },
                "nombrePrimerTrabajo": {
                    "type": "string"
                },
                "trabajaActualmente": {
                    "type": "boolean"
                },
                "trabajaSistemas": {
                    "type": "boolean"
                },
                "experienciaSistemas": {
                    "type": "string"
                },
                "estudiosSistemas": {
                    "type": "string"
                },
                "fechaActualTrabajo": {
                    "type": "string"
                },
                "lugarActualTrabajo": {
                    "type": "string"
                },
                "fechaPrimerEmpleo": {
                    "type": "string"
                },
                "topico": {
                    "type": "string"
                },
                "nivelInglesId": {
                    "type": "integer",
                    "$ref": "#/definitions/NivelIngles"
                },
                "estadoId": {
                    "type": "integer",
                    "$ref": "#/definitions/Estado"
                },
                "sedeId": {
                    "type": "integer",
                    "$ref": "#/definitions/Sede"
                },
                "nodoId": {
                    "type": "integer",
                    "$ref": "#/definitions/Nodo"
                }
            }
        },
...
        "Nodo": {
            "properties": {
                "id": {
                    "type": "integer",
                    "uniqueItems": true
                },
                "nombre": {
                    "type": "string"
                }
            }
        },
        "Sede": {
            "properties": {
                "id": {
                    "type": "integer",
                    "uniqueItems": true
                },
                "nombre": {
                    "type": "string"
                },
                "NodoId": {
                    "type": "integer",
                    "$ref": "#/definitions/Nodo"
                }
            }
        },
        "Curso": {
            "properties": {
                "id": {
                    "type": "integer",
                    "uniqueItems": true
                },
                "horario": {
                    "type": "string"
                },
                "profesores": {
                    "type": "string"
                },
                "notas": {
                    "type": "string"
                },
                "anio": {
                    "type": "integer"
                },
                "periodo": {
                    "type": "string",
                    "$ref": "#/definitions/Periodo"
                },
                "TopicoId": {
                    "type": "integer"
                },
                "estado": {
                    "type": "boolean"
                },
                "NodoId": {
                    "type": "integer",
                    "$ref": "#/definitions/Nodo"
                },
                "SedeId": {
                    "type": "integer",
                    "$ref": "#/definitions/Sede"
                },
                "topico": {
                    "type": "string"
                }
            }
        },
...
        "Topico": {
            "properties": {
                "id": {
                    "type": "integer",
                    "uniqueItems": true
                },
                "nombre": {
                    "type": "string"
                }
            }
        },
        "Periodo": {
            "properties": {
                "id": {
                    "type": "integer",
                    "uniqueItems": true
                },
                "anio": {
                    "type": "integer"
                },
                "periodo": {
                    "type": "string"
                },
                "TopicoId": {
                    "type": "integer",
                    "$ref": "#/definitions/Topico"
                },
                "estado": {
                    "type": "boolean"
                },
                "mensajeDeCierre": {
                    "type": "string"
                }
            }
        },
...
        "Estado": {
            "properties": {
                "id": {
                    "type": "integer",
                    "uniqueItems": true
                },
                "nombre": {
                    "type": "string"
                }
            }
        },
...
        "NivelIngles": {
            "properties": {
                "id": {
                    "type": "integer",
                    "uniqueItems": true
                },
                "nombre": {
                    "type": "string"
                }
            }
        },
...
        "DatoTopico": {
            "properties": {
                "id": {
                    "type": "integer",
                    "uniqueItems": true
                },
                "nombre": {
                    "type": "string"
                }
            }
        },
        "PostTopico": {
            "properties": {
                "nombre": {
                    "type": "string"
                }
            }
        },
        "CursoEditar": {
            "properties": {
                "NodoId": {
                    "type":"integer"
                },
                "SedeId": {
                    "type":"integer"
                },
                "horario": {
                    "type": "string"
                },
                "profesores": {
                    "type": "string"
                },
                "notas": {
                    "type": "string"
                },
                "anio": {
                    "type": "integer"
                },
                "periodo": {
                    "type": "integer"
                },
                "TopicoId": {
                    "type": "integer"
                },
                "estado": {
                    "type": "boolean"
                }
            }
        }
    }
}